requirements.txt
tweepy
textblob
python-telegram-bot==13.15
python-dotenv
